+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.5.0 (unreleased)
 ------------------
 
+- #2343 Allow to define the sorting criteria for Result Options
 - #2345 Retrieve setup items by object
 - #2344 Fix error on sample copy when fields are hidden
 - #2334 Remove legacy reports